+++ apps/client/instamint-webapp/src/pages/profile/settings/notifications/index.tsx
@@ -1,10 +1,9 @@
 import type { GetServerSideProps } from "next"
 import { serverSideTranslations } from "next-i18next/serverSideTranslations"
-import type { ReactElement } from "react"
-import React from "react"
+import React, { type ReactElement } from "react"
 
-import getTranslationBaseImports from "@/web/utils/helpers/getTranslationBaseImports"
 import SettingsLayout from "@/web/components/layout/SettingsLayout"
+import getTranslationBaseImports from "@/web/utils/helpers/getTranslationBaseImports"
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   const { locale } = context